pyproject.toml
[project]
name = "practiced-23june"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.3",
    "pydantic>=2.11.7",
    "python-decouple>=3.8",
]
